# apps/web/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Followw UnB API",
        "description": "Cliente alternativo e open-source para o ecossistema de sistemas da Universidade de Brasília (UnB). Agrega em uma experiência unificada e performática dados do SIGAA, Restaurante Universitário (RU) e calendário oficial, com autenticação stateless delegada ao CAS da UnB, credenciais cifradas em cookies JWE (AES-256-GCM), sincronização assíncrona inteligente e cache resiliente via Stale-While-Revalidate.",
        "version": "0.1.0"
    },
    "paths": {
...
        "/classrooms/{classroom_id}/news": {
            "get": {
                "tags": ["Classrooms"],
                "summary": "Consultar notícias de uma turma do usuário",
                "description": "ID, título e dia das notícias da turma. Cache de 60 minutos, atualizado em segundo plano após vencer; notícias removidas no SIGAA são preservadas.",
                "operationId": "get_classroom_news_classrooms__classroom_id__news_get",
                "parameters": [
                    {
                        "name": "classroom_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Classroom.id ou o classroom_sigaa_id de /news.",
                            "title": "Classroom Id"
                        },
                        "description": "Classroom.id ou o classroom_sigaa_id de /news."
                    },
                    {
                        "name": "Cache-Control",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [{ "type": "string" }, { "type": "null" }],
                            "description": "`no-cache` (ou `max-age=0`) busca na origem antes de responder; `max-age=N` aceita cache de até N s; `stale-if-error[=N]` devolve o cache se a origem falhar; `only-if-cached` nunca busca (504 sem cache).",
                            "title": "Cache-Control"
                        },
                        "description": "`no-cache` (ou `max-age=0`) busca na origem antes de responder; `max-age=N` aceita cache de até N s; `stale-if-error[=N]` devolve o cache se a origem falhar; `only-if-cached` nunca busca (504 sem cache)."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/News" },
                                    "title": "Response Get Classroom News Classrooms  Classroom Id  News Get"
                                }
                            }
                        }
                    },
                    "401": { "description": "Credenciais ausentes ou inválidas." },
                    "502": { "description": "SIGAA indisponível." },
                    "404": { "description": "Turma não encontrada entre as turmas do usuário." },
                    "503": { "description": "Cache em atualização." },
                    "504": { "description": "Notícias ainda não disponíveis no cache." },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            }
        },
        "/classrooms/{classroom_id}/news/{news_id}": {
            "get": {
                "tags": ["Classrooms"],
                "summary": "Consultar conteúdo e anexos de uma notícia da turma",
                "description": "Texto em Markdown, data e hora e anexos. Conteúdo salvo no primeiro acesso, sem revalidação automática. Cache-Control: no-cache força uma nova consulta.",
                "operationId": "get_classroom_news_detail_classrooms__classroom_id__news__news_id__get",
                "parameters": [
                    {
                        "name": "classroom_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Classroom.id ou o classroom_sigaa_id de /news.",
                            "title": "Classroom Id"
                        },
                        "description": "Classroom.id ou o classroom_sigaa_id de /news."
                    },
                    {
                        "name": "news_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "exclusiveMinimum": 0,
                            "description": "ID da notícia na listagem da turma.",
                            "title": "News Id"
                        },
                        "description": "ID da notícia na listagem da turma."
                    },
                    {
                        "name": "Cache-Control",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [{ "type": "string" }, { "type": "null" }],
                            "description": "`no-cache` (ou `max-age=0`) busca na origem antes de responder; `max-age=N` aceita cache de até N s; `stale-if-error[=N]` devolve o cache se a origem falhar; `only-if-cached` nunca busca (504 sem cache).",
                            "title": "Cache-Control"
                        },
                        "description": "`no-cache` (ou `max-age=0`) busca na origem antes de responder; `max-age=N` aceita cache de até N s; `stale-if-error[=N]` devolve o cache se a origem falhar; `only-if-cached` nunca busca (504 sem cache)."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/News" }
                            }
                        }
                    },
                    "401": { "description": "Credenciais ausentes ou inválidas." },
                    "502": { "description": "SIGAA indisponível." },
                    "404": {
                        "description": "Turma fora da lista do usuário ou notícia ausente na turma."
                    },
                    "503": { "description": "Cache em atualização." },
                    "504": { "description": "Conteúdo ainda não disponível no cache." },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            }
        },
        "/classrooms": {
            "get": {
                "tags": ["Classrooms"],
                "summary": "Consultar as turmas do usuário autenticado",
                "operationId": "get_classrooms_classrooms_get",
                "parameters": [
                    {
                        "name": "semester",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                { "type": "string", "pattern": "^(all|[0-9]{4}\\.[0-9])$" },
                                { "type": "null" }
                            ],
                            "description": "Sem filtro: turmas atuais. Use 'all' ou um semestre no formato AAAA.P, como 2026.2.",
                            "title": "Semester"
                        },
                        "description": "Sem filtro: turmas atuais. Use 'all' ou um semestre no formato AAAA.P, como 2026.2."
                    },
                    {
                        "name": "Cache-Control",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [{ "type": "string" }, { "type": "null" }],
                            "description": "`no-cache` (ou `max-age=0`) busca na origem antes de responder; `max-age=N` aceita cache de até N s; `stale-if-error[=N]` devolve o cache se a origem falhar; `only-if-cached` nunca busca (504 sem cache).",
                            "title": "Cache-Control"
                        },
                        "description": "`no-cache` (ou `max-age=0`) busca na origem antes de responder; `max-age=N` aceita cache de até N s; `stale-if-error[=N]` devolve o cache se a origem falhar; `only-if-cached` nunca busca (504 sem cache)."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/Classroom" },
                                    "title": "Response Get Classrooms Classrooms Get"
                                }
                            }
                        }
                    },
                    "401": { "description": "Credenciais ausentes ou inválidas." },
                    "502": { "description": "SIGAA indisponível." },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            }
        },
        "/classrooms/frequency": {
            "get": {
                "tags": ["Classrooms"],
                "summary": "Consultar frequência de todas as turmas atuais",
                "description": "Turmas atuais com identificação, andamento, frequência, frequency_status e resumo das entradas. Falha em uma turma retorna erro, sem omiti-la da lista.",
                "operationId": "get_current_frequencies_classrooms_frequency_get",
                "parameters": [
                    {
                        "name": "Cache-Control",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [{ "type": "string" }, { "type": "null" }],
                            "description": "`no-cache` (ou `max-age=0`) busca na origem antes de responder; `max-age=N` aceita cache de até N s; `stale-if-error[=N]` devolve o cache se a origem falhar; `only-if-cached` nunca busca (504 sem cache).",
                            "title": "Cache-Control"
                        },
                        "description": "`no-cache` (ou `max-age=0`) busca na origem antes de responder; `max-age=N` aceita cache de até N s; `stale-if-error[=N]` devolve o cache se a origem falhar; `only-if-cached` nunca busca (504 sem cache)."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClassroomFrequencyResult"
                                    },
                                    "title": "Response Get Current Frequencies Classrooms Frequency Get"
                                }
                            }
                        }
                    },
                    "401": { "description": "Credenciais ausentes ou inválidas." },
                    "502": { "description": "SIGAA indisponível." },
                    "404": { "description": "Turma não encontrada entre as turmas do usuário." },
                    "503": { "description": "Cache em atualização." },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            }
        },
        "/classrooms/{classroom_id}/frequency": {
            "get": {
                "tags": ["Classrooms"],
                "summary": "Consultar frequência e andamento de uma turma",
                "description": "Aceita Classroom.id (hash) ou sigaa_id numérico. frequency_status indica not_registered, partially_registered ou registered nas entradas publicadas.",
                "operationId": "get_classroom_frequency_classrooms__classroom_id__frequency_get",
                "parameters": [
                    {
                        "name": "classroom_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Classroom.id ou Classroom.sigaa_id.",
                            "title": "Classroom Id"
                        },
                        "description": "Classroom.id ou Classroom.sigaa_id."
                    },
                    {
                        "name": "Cache-Control",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [{ "type": "string" }, { "type": "null" }],
                            "description": "`no-cache` (ou `max-age=0`) busca na origem antes de responder; `max-age=N` aceita cache de até N s; `stale-if-error[=N]` devolve o cache se a origem falhar; `only-if-cached` nunca busca (504 sem cache).",
                            "title": "Cache-Control"
                        },
                        "description": "`no-cache` (ou `max-age=0`) busca na origem antes de responder; `max-age=N` aceita cache de até N s; `stale-if-error[=N]` devolve o cache se a origem falhar; `only-if-cached` nunca busca (504 sem cache)."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/ClassroomFrequency" }
                            }
                        }
                    },
                    "401": { "description": "Credenciais ausentes ou inválidas." },
                    "502": { "description": "SIGAA indisponível." },
                    "404": { "description": "Turma não encontrada entre as turmas do usuário." },
                    "503": { "description": "Cache em atualização." },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            }
        },
        "/classrooms/{classroom_id}/members": {
            "get": {
                "tags": ["Classrooms"],
                "summary": "Consultar docentes e discentes de uma turma do usuário",
                "operationId": "get_classroom_members_classrooms__classroom_id__members_get",
                "parameters": [
                    {
                        "name": "classroom_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Classroom.id ou Classroom.sigaa_id.",
                            "title": "Classroom Id"
                        },
                        "description": "Classroom.id ou Classroom.sigaa_id."
                    },
                    {
                        "name": "Cache-Control",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [{ "type": "string" }, { "type": "null" }],
                            "description": "`no-cache` (ou `max-age=0`) busca na origem antes de responder; `max-age=N` aceita cache de até N s; `stale-if-error[=N]` devolve o cache se a origem falhar; `only-if-cached` nunca busca (504 sem cache).",
                            "title": "Cache-Control"
                        },
                        "description": "`no-cache` (ou `max-age=0`) busca na origem antes de responder; `max-age=N` aceita cache de até N s; `stale-if-error[=N]` devolve o cache se a origem falhar; `only-if-cached` nunca busca (504 sem cache)."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/ClassroomMember" },
                                    "title": "Response Get Classroom Members Classrooms  Classroom Id  Members Get"
                                }
                            }
                        }
                    },
                    "401": { "description": "Credenciais ausentes ou inválidas." },
                    "502": { "description": "SIGAA indisponível." },
                    "404": { "description": "Turma não encontrada entre as turmas do usuário." },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            }
        },
        "/classrooms/{classroom_id}/statistics": {
            "get": {
                "tags": ["Classrooms"],
                "summary": "Consultar a situação dos discentes de uma turma do usuário",
                "operationId": "get_classroom_statistics_classrooms__classroom_id__statistics_get",
                "parameters": [
                    {
                        "name": "classroom_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Classroom.id ou Classroom.sigaa_id.",
                            "title": "Classroom Id"
                        },
                        "description": "Classroom.id ou Classroom.sigaa_id."
                    },
                    {
                        "name": "Cache-Control",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [{ "type": "string" }, { "type": "null" }],
                            "description": "`no-cache` (ou `max-age=0`) busca na origem antes de responder; `max-age=N` aceita cache de até N s; `stale-if-error[=N]` devolve o cache se a origem falhar; `only-if-cached` nunca busca (504 sem cache).",
                            "title": "Cache-Control"
                        },
                        "description": "`no-cache` (ou `max-age=0`) busca na origem antes de responder; `max-age=N` aceita cache de até N s; `stale-if-error[=N]` devolve o cache se a origem falhar; `only-if-cached` nunca busca (504 sem cache)."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/StatisticsShare" },
                                    "title": "Response Get Classroom Statistics Classrooms  Classroom Id  Statistics Get"
                                }
                            }
                        }
                    },
                    "401": { "description": "Credenciais ausentes ou inválidas." },
                    "502": { "description": "SIGAA indisponível." },
                    "404": { "description": "Turma não encontrada entre as turmas do usuário." },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Classroom": {
                "properties": {
                    "id": { "type": "string", "title": "Id" },
                    "sigaa_id": {
                        "anyOf": [{ "type": "integer" }, { "type": "null" }],
                        "title": "Sigaa Id"
                    },
                    "number": { "type": "string", "title": "Number" },
                    "semester": { "type": "string", "title": "Semester" },
                    "schedule": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Schedule"
                    },
                    "room": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Room"
                    },
                    "current": { "type": "boolean", "title": "Current", "default": false },
                    "subject": { "$ref": "#/components/schemas/Subject" }
                },
                "type": "object",
                "required": ["id", "number", "semester", "subject"],
                "title": "Classroom",
                "description": "Turma, nos campos da tabela `classrooms`.\n\n`id` é o `frontEndIdTurma` do SIGAA — o único identificador presente tanto\nnas turmas do semestre quanto no histórico. `sigaa_id` é o id numérico\ninterno, que só o portal do semestre expõe. `current` marca as turmas que o\nportal lista como atuais."
            },
...
            "ClassroomFrequency": {
                "properties": {
                    "progress": { "$ref": "#/components/schemas/ClassroomProgress" },
                    "frequency": {
                        "anyOf": [
                            { "$ref": "#/components/schemas/ClassroomAttendance" },
                            { "type": "null" }
                        ]
                    },
                    "frequency_status": {
                        "type": "string",
                        "enum": ["not_registered", "partially_registered", "registered"],
                        "title": "Frequency Status",
                        "readOnly": true
                    }
                },
                "type": "object",
                "required": ["progress", "frequency_status"],
                "title": "ClassroomFrequency",
                "description": "A tela de frequência da turma virtual.\n\n`frequency` é `None` quando o docente não lançou frequência — nesse caso o\nSIGAA ainda mostra totais na tela, mas eles são fictícios (100% de presença\nem toda a carga horária), então não são devolvidos."
            },
            "ClassroomFrequencyResult": {
                "properties": {
                    "progress": { "$ref": "#/components/schemas/ClassroomProgress" },
                    "frequency": {
                        "anyOf": [
                            { "$ref": "#/components/schemas/ClassroomAttendance" },
                            { "type": "null" }
                        ]
                    },
                    "classroom": { "$ref": "#/components/schemas/Classroom" },
                    "frequency_status": {
                        "type": "string",
                        "enum": ["not_registered", "partially_registered", "registered"],
                        "title": "Frequency Status",
                        "readOnly": true
                    }
                },
                "type": "object",
                "required": ["progress", "classroom", "frequency_status"],
                "title": "ClassroomFrequencyResult"
            },
            "ClassroomMember": {
                "properties": {
                    "name": { "type": "string", "title": "Name" },
                    "role": { "$ref": "#/components/schemas/ClassroomRole" },
                    "registration": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Registration"
                    },
                    "photo": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Photo"
                    },
                    "email": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Email"
                    },
                    "course": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Course"
                    },
                    "unity": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Unity"
                    },
                    "person_id": {
                        "anyOf": [{ "type": "integer" }, { "type": "null" }],
                        "title": "Person Id"
                    }
                },
                "type": "object",
                "required": ["name", "role"],
                "title": "ClassroomMember",
                "description": "Participante de uma turma, nos campos da tabela `users` que a tela expõe."
            },
...
            "ClassroomRole": {
                "type": "string",
                "enum": ["aluno", "professor", "monitor"],
                "title": "ClassroomRole"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": { "$ref": "#/components/schemas/ValidationError" },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "News": {
                "properties": {
                    "id": { "anyOf": [{ "type": "integer" }, { "type": "null" }], "title": "Id" },
                    "classroom_sigaa_id": {
                        "anyOf": [{ "type": "integer" }, { "type": "null" }],
                        "title": "Classroom Sigaa Id"
                    },
                    "title": { "type": "string", "title": "Title" },
                    "published_on": { "type": "string", "format": "date", "title": "Published On" },
                    "published_at": {
                        "anyOf": [{ "type": "string", "format": "date-time" }, { "type": "null" }],
                        "title": "Published At"
                    },
                    "content": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Content"
                    },
                    "attachments": {
                        "items": { "$ref": "#/components/schemas/NewsAttachment" },
                        "type": "array",
                        "title": "Attachments",
                        "default": []
                    }
                },
                "type": "object",
                "required": ["title", "published_on"],
                "title": "News",
                "description": "A home não traz `id`; hora, texto e anexos só vêm de `get_classroom_news`."
            },
            "NewsAttachment": {
                "properties": {
                    "name": { "type": "string", "title": "Name" },
                    "url": { "type": "string", "title": "Url" }
                },
                "type": "object",
                "required": ["name", "url"],
                "title": "NewsAttachment"
            },
...
            "StatisticsShare": {
                "properties": {
                    "situation": { "$ref": "#/components/schemas/StudentSituation" },
                    "percentage": { "type": "number", "title": "Percentage" }
                },
                "type": "object",
                "required": ["situation", "percentage"],
                "title": "StatisticsShare",
                "description": "Uma fatia do gráfico de estatísticas: a porcentagem dos discentes da turma.\n\nO SIGAA arredonda cada fatia em uma casa, então a soma pode fechar em 99.9\nou 100.1. O gráfico não expõe a contagem de alunos, só a porcentagem."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": { "type": "string", "title": "Message" },
                    "type": { "type": "string", "title": "Error Type" },
                    "input": { "title": "Input" },
                    "ctx": { "type": "object", "title": "Context" }
                },
                "type": "object",
                "required": ["loc", "msg", "type"],
                "title": "ValidationError"
            }
        }
    },
...
}